Simplify discord bot and remove unused modules
