--- discord-bot/src/utils/components.js
@@ -1,13 +1,0 @@
-const { ActionRowBuilder, ButtonBuilder, ButtonStyle } = require('discord.js');
-
-function createBackToTownRow() {
-  return new ActionRowBuilder().addComponents(
-    new ButtonBuilder()
-      .setCustomId('nav-town')
-      .setLabel('Back to Town')
-      .setStyle(ButtonStyle.Secondary)
-      .setEmoji('🏠')
-  );
-}
-
-module.exports = { createBackToTownRow };
