--- discord-bot/tests/admin.test.js
@@ -1,291 +1,0 @@
-const admin = require('../commands/admin');
-
-jest.mock('../src/utils/userService', () => ({
-  getUser: jest.fn(),
-  addAbility: jest.fn()
-}));
-jest.mock('../src/utils/embedBuilder', () => ({
-  sendCardDM: jest.fn(),
-  sendWeaponDM: jest.fn(),
-  sendItemDM: jest.fn()
-}));
-jest.mock('../src/utils/weaponService', () => ({
-  addWeapon: jest.fn()
-}));
-jest.mock('../src/utils/armorService', () => ({
-  addArmor: jest.fn()
-}));
-
-const userService = require('../src/utils/userService');
-const { sendCardDM, sendWeaponDM, sendItemDM } = require('../src/utils/embedBuilder');
-const weaponService = require('../src/utils/weaponService');
-const armorService = require('../src/utils/armorService');
-const gameData = require('../util/gameData');
-const { allPossibleAbilities, allPossibleWeapons, allPossibleArmors } = require('../../backend/game/data');
-
-function createInteraction(role = 'Game Master', sub = 'grant-ability') {
-  return {
-    member: { roles: { cache: { some: jest.fn(fn => fn({ name: role })) } } },
-    options: {
-      getSubcommand: jest.fn().mockReturnValue(sub),
-      getUser: jest.fn().mockReturnValue({ id: '200', username: 'Target' }),
-      getString: jest.fn()
-    },
-    reply: jest.fn().mockResolvedValue(),
-    followUp: jest.fn().mockResolvedValue()
-  };
-}
-
-describe('admin grant-ability command', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-    gameData.gameData.abilities = new Map(allPossibleAbilities.map(a => [a.id, a]));
-  });
-
-  test('requires Game Master role', async () => {
-    const interaction = createInteraction('Player');
-    interaction.options.getString.mockReturnValue('Power Strike');
-    await admin.execute(interaction);
-    expect(interaction.reply).toHaveBeenCalledWith(expect.objectContaining({
-      ephemeral: true,
-      content: expect.stringContaining('necessary permissions')
-    }));
-  });
-
-  test('errors when ability not found', async () => {
-    const interaction = createInteraction();
-    interaction.options.getString.mockReturnValue('Nonexistent');
-    userService.getUser.mockResolvedValue({ id: 1 });
-    await admin.execute(interaction);
-    expect(interaction.reply).toHaveBeenCalledWith(expect.objectContaining({
-      ephemeral: true,
-      content: expect.stringContaining('Could not find an ability')
-    }));
-  });
-
-  test('errors when user not found', async () => {
-    const interaction = createInteraction();
-    interaction.options.getString.mockReturnValue('Power Strike');
-    userService.getUser.mockResolvedValue(null);
-    await admin.execute(interaction);
-    expect(interaction.reply).toHaveBeenCalledWith(expect.objectContaining({
-      ephemeral: true,
-      content: expect.stringContaining('has not started playing yet')
-    }));
-  });
-
-  test('grants ability and sends DM', async () => {
-    const interaction = createInteraction();
-    interaction.options.getString.mockReturnValue('Power Strike');
-    userService.getUser.mockResolvedValue({ id: 1 });
-    userService.addAbility.mockResolvedValue(99);
-    await admin.execute(interaction);
-    expect(userService.addAbility).toHaveBeenCalled();
-    expect(sendCardDM).toHaveBeenCalled();
-    expect(interaction.reply).toHaveBeenCalledWith(expect.objectContaining({
-      ephemeral: true,
-      content: expect.stringContaining('successfully granted')
-    }));
-  });
-
-  test('notifies when ability card DM fails', async () => {
-    const interaction = createInteraction();
-    interaction.options.getString.mockReturnValue('Power Strike');
-    userService.getUser.mockResolvedValue({ id: 1 });
-    userService.addAbility.mockResolvedValue(99);
-    sendCardDM.mockRejectedValue(new Error('fail'));
-    await admin.execute(interaction);
-    expect(sendCardDM).toHaveBeenCalled();
-    expect(interaction.followUp).toHaveBeenCalledWith(
-      expect.objectContaining({ ephemeral: true })
-    );
-  });
-
-  test('autocomplete suggests ability names', async () => {
-    const interaction = {
-      options: { getFocused: jest.fn().mockReturnValue({ name: 'ability', value: 'Power' }) },
-      respond: jest.fn().mockResolvedValue()
-    };
-    await admin.autocomplete(interaction);
-    const options = interaction.respond.mock.calls[0][0];
-    expect(options.some(o => o.name === 'Power Strike')).toBe(true);
-  });
-
-  test('autocomplete does not suggest nonexistent abilities', async () => {
-    const interaction = {
-      options: { getFocused: jest.fn().mockReturnValue({ name: 'ability', value: 'Nonexistent' }) },
-      respond: jest.fn().mockResolvedValue()
-    };
-    await admin.autocomplete(interaction);
-    const options = interaction.respond.mock.calls[0][0];
-    expect(options).toHaveLength(0);
-  });
-});
-
-describe('admin grant-weapon command', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-
-  test('requires Game Master role', async () => {
-    const interaction = createInteraction('Player', 'grant-weapon');
-    interaction.options.getString.mockReturnValue('Worn Sword');
-    await admin.execute(interaction);
-    expect(interaction.reply).toHaveBeenCalledWith(expect.objectContaining({
-      ephemeral: true,
-      content: expect.stringContaining('necessary permissions')
-    }));
-  });
-
-  test('errors when weapon not found', async () => {
-    const interaction = createInteraction('Game Master', 'grant-weapon');
-    interaction.options.getString.mockReturnValue('Nonexistent');
-    await admin.execute(interaction);
-    expect(interaction.reply).toHaveBeenCalledWith(expect.objectContaining({
-      ephemeral: true,
-      content: expect.stringContaining('Could not find a weapon')
-    }));
-  });
-
-  test('errors when user not found', async () => {
-    const interaction = createInteraction('Game Master', 'grant-weapon');
-    interaction.options.getString.mockReturnValue('Worn Sword');
-    userService.getUser.mockResolvedValue(null);
-    await admin.execute(interaction);
-    expect(interaction.reply).toHaveBeenCalledWith(expect.objectContaining({
-      ephemeral: true,
-      content: expect.stringContaining('has not started playing yet')
-    }));
-  });
-
-  test('grants weapon and sends DM', async () => {
-    const interaction = createInteraction('Game Master', 'grant-weapon');
-    interaction.options.getString.mockReturnValue('Worn Sword');
-    userService.getUser.mockResolvedValue({ id: 1 });
-    weaponService.addWeapon.mockResolvedValue(99);
-    await admin.execute(interaction);
-    expect(weaponService.addWeapon).toHaveBeenCalled();
-    expect(sendWeaponDM).toHaveBeenCalled();
-    expect(interaction.reply).toHaveBeenCalledWith(expect.objectContaining({
-      ephemeral: true,
-      content: expect.stringContaining('successfully granted')
-    }));
-  });
-
-  test('notifies when weapon DM fails', async () => {
-    const interaction = createInteraction('Game Master', 'grant-weapon');
-    interaction.options.getString.mockReturnValue('Worn Sword');
-    userService.getUser.mockResolvedValue({ id: 1 });
-    weaponService.addWeapon.mockResolvedValue(99);
-    sendWeaponDM.mockRejectedValue(new Error('fail'));
-    await admin.execute(interaction);
-    expect(sendWeaponDM).toHaveBeenCalled();
-    expect(interaction.followUp).toHaveBeenCalledWith(
-      expect.objectContaining({ ephemeral: true })
-    );
-  });
-
-  test("autocomplete suggests weapon names", async () => {
-    const interaction = {
-      options: { getFocused: jest.fn().mockReturnValue({ name: 'weapon', value: 'Worn' }) },
-      respond: jest.fn().mockResolvedValue()
-    };
-    await admin.autocomplete(interaction);
-    const options = interaction.respond.mock.calls[0][0];
-    expect(options.some(o => o.name.includes("Worn"))).toBe(true);
-  });
-
-  test("autocomplete does not suggest nonexistent weapons", async () => {
-    const interaction = {
-      options: { getFocused: jest.fn().mockReturnValue({ name: 'weapon', value: 'Nonexistent' }) },
-      respond: jest.fn().mockResolvedValue()
-    };
-    await admin.autocomplete(interaction);
-    const options = interaction.respond.mock.calls[0][0];
-    expect(options).toHaveLength(0);
-  });
-});
-
-describe('admin grant-armor command', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-
-  test('requires Game Master role', async () => {
-    const interaction = createInteraction('Player', 'grant-armor');
-    interaction.options.getString.mockReturnValue('Leather Padding');
-    await admin.execute(interaction);
-    expect(interaction.reply).toHaveBeenCalledWith(expect.objectContaining({
-      ephemeral: true,
-      content: expect.stringContaining('necessary permissions')
-    }));
-  });
-
-  test('errors when armor not found', async () => {
-    const interaction = createInteraction('Game Master', 'grant-armor');
-    interaction.options.getString.mockReturnValue('Nonexistent');
-    await admin.execute(interaction);
-    expect(interaction.reply).toHaveBeenCalledWith(expect.objectContaining({
-      ephemeral: true,
-      content: expect.stringContaining('Could not find an armor')
-    }));
-  });
-
-  test('errors when user not found', async () => {
-    const interaction = createInteraction('Game Master', 'grant-armor');
-    interaction.options.getString.mockReturnValue('Leather Padding');
-    userService.getUser.mockResolvedValue(null);
-    await admin.execute(interaction);
-    expect(interaction.reply).toHaveBeenCalledWith(expect.objectContaining({
-      ephemeral: true,
-      content: expect.stringContaining('has not started playing yet')
-    }));
-  });
-
-  test('grants armor and sends DM', async () => {
-    const interaction = createInteraction('Game Master', 'grant-armor');
-    interaction.options.getString.mockReturnValue('Leather Padding');
-    userService.getUser.mockResolvedValue({ id: 1 });
-    armorService.addArmor.mockResolvedValue(99);
-    await admin.execute(interaction);
-    expect(armorService.addArmor).toHaveBeenCalled();
-    expect(sendItemDM).toHaveBeenCalled();
-    expect(interaction.reply).toHaveBeenCalledWith(expect.objectContaining({
-      ephemeral: true,
-      content: expect.stringContaining('successfully granted')
-    }));
-  });
-
-  test('notifies when armor DM fails', async () => {
-    const interaction = createInteraction('Game Master', 'grant-armor');
-    interaction.options.getString.mockReturnValue('Leather Padding');
-    userService.getUser.mockResolvedValue({ id: 1 });
-    armorService.addArmor.mockResolvedValue(99);
-    sendItemDM.mockRejectedValue(new Error('fail'));
-    await admin.execute(interaction);
-    expect(sendItemDM).toHaveBeenCalled();
-    expect(interaction.followUp).toHaveBeenCalledWith(
-      expect.objectContaining({ ephemeral: true })
-    );
-  });
-
-  test('autocomplete suggests armor names', async () => {
-    const interaction = {
-      options: { getFocused: jest.fn().mockReturnValue({ name: 'armor', value: 'Leather' }) },
-      respond: jest.fn().mockResolvedValue()
-    };
-    await admin.autocomplete(interaction);
-    const options = interaction.respond.mock.calls[0][0];
-    expect(options.some(o => o.name.includes('Leather'))).toBe(true);
-  });
-
-  test('autocomplete does not suggest nonexistent armors', async () => {
-    const interaction = {
-      options: { getFocused: jest.fn().mockReturnValue({ name: 'armor', value: 'Nonexistent' }) },
-      respond: jest.fn().mockResolvedValue()
-    };
-    await admin.autocomplete(interaction);
-    const options = interaction.respond.mock.calls[0][0];
-    expect(options).toHaveLength(0);
-  });
-});
