--- discord-bot/tests/town.test.js
@@ -1,14 +1,0 @@
-const town = require('../commands/town');
-
-
-jest.mock('../src/utils/userService', () => ({ setUserLocation: jest.fn() }));
-const userService = require('../src/utils/userService');
-
-test('replies with town hub embed', async () => {
-  const interaction = { user: { id: '1' }, reply: jest.fn().mockResolvedValue() };
-  await town.execute(interaction);
-  expect(userService.setUserLocation).toHaveBeenCalledWith('1', 'town');
-  expect(interaction.reply).toHaveBeenCalledWith(
-    expect.objectContaining({ embeds: expect.any(Array), components: expect.any(Array), ephemeral: true })
-  );
-});
